+++ test/hexapdf/type/acro_form/test_variable_text_field.rb
@@ -108,6 +108,11 @@ describe HexaPDF::Type::AcroForm::VariableTextField do
                    @field.parse_default_appearance_string)
     end
 
+    it "converts the /DA to a string in case an invalid PDF uses a Symbol" do
+      @field[:DA] = :"1 g /F1 20 Tf"
+      assert_equal([:F1, 20, @color], @field.parse_default_appearance_string)
+    end
+
     it "fails if no /DA value is set and no default appearance string should be set" do
       @doc.acro_form.delete(:DA)
       @doc.config['acro_form.fallback_default_appearance'] = nil